+++ docs/News.rst
@@ -16,6 +16,8 @@ Features
   json.dumps/loads. A subclass of StringCol. Requires ``VARCHAR``/``TEXT``
   columns at backends, doesn't work with ``JSON`` columns.
 
+* Extend/fix support for ``DateTime`` from ``Zope``.
+
 * Drop support for very old version of ``mxDateTime``
   without ``mx.`` namespace.